+++ src/platform/dart/executor/task.rs
@@ -2,15 +2,23 @@
 //!
 //! [`platform::dart::executor`]: crate::platform::executor
 
+#[cfg(doc)]
+use std::future::Future;
 use std::{
-    cell::{Cell, RefCell},
-    mem::ManuallyDrop,
-    rc::Rc,
-    task::{Context, Poll, RawWaker, RawWakerVTable, Waker},
+    cell::RefCell,
+    sync::{
+        atomic::{AtomicBool, Ordering},
+        Arc,
+    },
+    task::{Context, Waker},
+    thread::{self, ThreadId},
 };
 
 use derive_more::Debug;
-use futures::future::LocalBoxFuture;
+use futures::{
+    future::LocalBoxFuture,
+    task::{self, ArcWake},
+};
 
 use crate::platform::dart::executor::task_wake;
 
@@ -41,21 +49,48 @@ pub struct Task {
 
     /// Indicates whether there is a [`Poll::Pending`] awake request of this
     /// [`Task`].
-    is_scheduled: Cell<bool>,
+    is_scheduled: AtomicBool,
+
+    /// ID of the thread this [`Task`] was created on and must be polled on.
+    thread_id: ThreadId,
+}
+
+/// [`Task`] can be sent across threads safely because it ensures that the
+/// underlying [`Future`] will only be touched from a single thread it was
+/// created on.
+unsafe impl Send for Task {}
+/// [`Task`] can be shared across threads safely because it ensures that the
+/// underlying [`Future`] will only be touched from a single thread it was
+/// created on.
+unsafe impl Sync for Task {}
+
+impl ArcWake for Task {
+    /// Commands an external Dart executor to poll this [`Task`] if it's
+    /// incomplete and there are no [`Poll::Pending`] awake requests already.
+    ///
+    /// [`Poll::Pending`]: task::Poll::Pending
+    fn wake_by_ref(arc_self: &Arc<Self>) {
+        if !arc_self.is_scheduled.swap(true, Ordering::AcqRel) {
+            task_wake(Arc::clone(arc_self));
+        }
+    }
 }
 
 impl Task {
     /// Spawns a new [`Task`] that will drive the given [`Future`].
     ///
+    /// Must be called on the same thread where the [`Task`] will be polled,
+    /// otherwise polling will panic.
+    ///
     /// [`Future`]: std::future::Future
     pub fn spawn(future: LocalBoxFuture<'static, ()>) {
-        let this = Rc::new(Self {
+        let this = Arc::new(Self {
             inner: RefCell::new(None),
-            is_scheduled: Cell::new(false),
+            is_scheduled: AtomicBool::new(false),
+            thread_id: thread::current().id(),
         });
 
-        let waker =
-            unsafe { Waker::from_raw(Self::into_raw_waker(Rc::clone(&this))) };
+        let waker = task::waker(Arc::clone(&this));
         drop(this.inner.borrow_mut().replace(Inner { future, waker }));
 
         Self::wake_by_ref(&this);
@@ -65,74 +100,37 @@ impl Task {
     ///
     /// Polling after [`Future`]'s completion is no-op.
     ///
+    /// # Panics
+    ///
+    /// If called not on the same thread where this [`Task`] was originally
+    /// created.
+    ///
     /// [`Future`]: std::future::Future
-    pub fn poll(&self) -> Poll<()> {
+    pub fn poll(&self) {
+        assert_eq!(
+            self.thread_id,
+            thread::current().id(),
+            "`dart::executor::Task` can only be polled on the same thread \
+             where it was originally created",
+        );
+
         let mut borrow = self.inner.borrow_mut();
 
         // Just ignore poll request if the `Future` is completed.
         let Some(inner) = borrow.as_mut() else {
-            return Poll::Ready(());
+            return;
         };
+
+        self.is_scheduled.store(false, Ordering::Release);
 
         let poll = {
             let mut cx = Context::from_waker(&inner.waker);
             inner.future.as_mut().poll(&mut cx)
         };
-        self.is_scheduled.set(false);
 
         // Cleanup resources if future is ready.
         if poll.is_ready() {
             *borrow = None;
         }
-
-        poll
-    }
-
-    /// Calls the [`task_wake()`] function by the provided reference if this
-    /// [`Task`] s incomplete and there are no [`Poll::Pending`] awake requests
-    /// already.
-    fn wake_by_ref(this: &Rc<Self>) {
-        if !this.is_scheduled.replace(true) {
-            task_wake(Rc::clone(this));
-        }
-    }
-
-    /// Pretty much a copy of [`std::task::Wake`] implementation but for
-    /// `Rc<?Send + ?Sync>` instead of `Arc<Send + Sync>` since we are sure
-    /// that everything will run on a single thread.
-    fn into_raw_waker(this: Rc<Self>) -> RawWaker {
-        #![expect( // not visible
-            clippy::missing_docs_in_private_items,
-            reason = "not visible at all"
-        )]
-
-        // Refer to `RawWakerVTable::new()` documentation for better
-        // understanding of what the following functions do.
-
-        unsafe fn raw_clone(ptr: *const ()) -> RawWaker {
-            let ptr =
-                ManuallyDrop::new(unsafe { Rc::from_raw(ptr.cast::<Task>()) });
-            Task::into_raw_waker(Rc::clone(&(*ptr)))
-        }
-
-        unsafe fn raw_wake(ptr: *const ()) {
-            let ptr = unsafe { Rc::from_raw(ptr.cast::<Task>()) };
-            Task::wake_by_ref(&ptr);
-        }
-
-        unsafe fn raw_wake_by_ref(ptr: *const ()) {
-            let ptr =
-                ManuallyDrop::new(unsafe { Rc::from_raw(ptr.cast::<Task>()) });
-            Task::wake_by_ref(&ptr);
-        }
-
-        unsafe fn raw_drop(ptr: *const ()) {
-            drop(unsafe { Rc::from_raw(ptr.cast::<Task>()) });
-        }
-
-        const VTABLE: RawWakerVTable =
-            RawWakerVTable::new(raw_clone, raw_wake, raw_wake_by_ref, raw_drop);
-
-        RawWaker::new(Rc::into_raw(this).cast::<()>(), &VTABLE)
     }
 }