+++ flutter/CHANGELOG.md
@@ -6,10 +6,14 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
-## [master] · unreleased
-[master]: /../../tree/master/flutter
+## [0.7.1] · 2024-12-30
+[0.7.1]: /../../tree/medea-jason-0.7.1/flutter
 
-See also [`medea-jason` crate `master` changes](/../../tree/master/CHANGELOG.md).
+See also [`medea-jason` crate 0.7.1 changes](/../../tree/medea-jason-0.7.1/CHANGELOG.md).
+
+### Fixed
+
+- Rust futures executor being called from non-main thread. ([#197])
 
 ### Upgraded
 
@@ -17,6 +21,7 @@ See also [`medea-jason` crate `master` changes](/../../tree/master/CHANGELOG.md)
     - [`flutter_rust_bridge`] to 2.7.0 version. ([#195])
 
 [#195]: /../../pull/195
+[#197]: /../../pull/197